requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.20.0
pyspark>=3.1.2
great-expectations>=0.14.0
langchain>=0.0.267
langchain-openai>=0.0.1
langchain-community>=0.0.1

# LLM integrations
openai>=0.28.0
transformers>=4.30.0
sentence-transformers>=2.2.2
torch>=1.10.0
huggingface-hub>=0.17.0

# Data storage & processing
pyarrow>=7.0.0
deltalake>=0.6.3
boto3>=1.20.0  # For S3 integration
sqlalchemy>=1.4.0  # For database connections

# Visualization & dashboard
streamlit>=1.15.0
plotly>=5.5.0
matplotlib>=3.5.0
seaborn>=0.12.0
pydeck>=0.7.0  # For geospatial visualizations

# Workflow orchestration
apache-airflow>=2.3.0

# API
fastapi>=0.90.0
uvicorn>=0.18.0

# Alerting
slackclient>=2.9.0
sendgrid>=6.9.0  # For email alerts

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Development
black>=22.3.0
flake8>=4.0.0
isort>=5.10.0
pre-commit>=2.20.0
